Checking for images before trying to foreach() over them in visual captcha's runonce.

+++ start.php
@@ -66,8 +66,10 @@ function visual_captcha_runonce() {
 		'limit' => 100
 	));
 
-	foreach ($images as $image) {
-		$image->delete();
+	if ($images) {
+		foreach ($images as $image) {
+			$image->delete();
+		}
 	}
 
 	// register an image
